fix:fixed bugs

+++ src/main/java/br/com/compass/challenge2/repositories/ProductRepository.java
@@ -1,9 +1,8 @@
 package br.com.compass.challenge2.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import br.com.compass.challenge2.model.Product;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
 }
